requirements.txt
python-telegram-bot
requests
urllib3
feedparser